backend/foodgram/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser
pillow
django-filter==23.5
reportlab==4.0.7
psycopg2-binary==2.9.3